requirements.txt
adlfs
pandas
azure-data-tables
Flask
tqdm
Flask-Bootstrap
pyarrow
azure-cognitiveservices-vision-computervision
azure-ai-vision
requests
numpy
Pillow